workflows: add pytest vcr for tests
