+++ cap/config.py
@@ -772,11 +772,12 @@ GITLAB_OAUTH_ACCESS_TOKEN = os.environ.get('APP_GITLAB_OAUTH_ACCESS_TOKEN')
 
 # Reana server url
 # ================
+REANA_SERVER_URL = os.environ.get('APP_REANA_SERVER_URL', 'https://reana.cern.ch')
 REANA_ACCESS_TOKEN = {
-    'ATLAS': os.environ.get('APP_REANA_ATLAS_ACCESS_TOKEN'),
-    'ALICE': os.environ.get('APP_REANA_ALICE_ACCESS_TOKEN'),
-    'CMS': os.environ.get('APP_REANA_CMS_ACCESS_TOKEN'),
-    'LHCb': os.environ.get('APP_REANA_LHCb_ACCESS_TOKEN')
+    'ATLAS': os.environ.get('APP_REANA_ATLAS_ACCESS_TOKEN', "default"),
+    'ALICE': os.environ.get('APP_REANA_ALICE_ACCESS_TOKEN', "default"),
+    'CMS': os.environ.get('APP_REANA_CMS_ACCESS_TOKEN', "default"),
+    'LHCb': os.environ.get('APP_REANA_LHCb_ACCESS_TOKEN', "default")
 }
 
 # Keytabs
