+++ tests/integration/workflows/cassettes/test_reana_workflow_logs.yaml
@@ -1,0 +1,51 @@
+interactions:
+- request:
+    body: null
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+    method: GET
+    uri: https://reana.cern.ch/api/workflows/4e4e6afc-fd57-4f59-836a-40ae19d2d6d0/logs
+  response:
+    body:
+      string: "{\n  \"logs\": \"{\\\"workflow_logs\\\": \\\"\\\", \\\"job_logs\\\"\
+        : {}, \\\"engine_specific\\\": null}\", \n  \"user\": \"63449ef7-be3a-4e77-b503-fb18a2e8a702\"\
+        , \n  \"workflow_id\": \"4e4e6afc-fd57-4f59-836a-40ae19d2d6d0\", \n  \"workflow_name\"\
+        : \"sturdy-wallaby.1\"\n}\n"
+    headers:
+      Access-Control-Allow-Origin:
+      - '*'
+      Access-Control-Expose-Headers:
+      - Content-Type, ETag, Link, X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset
+      Content-Length:
+      - '236'
+      Content-Type:
+      - application/json
+      Date:
+      - Mon, 14 Mar 2022 14:06:15 GMT
+      Referrer-Policy:
+      - strict-origin-when-cross-origin
+      Retry-After:
+      - '1'
+      Strict-Transport-Security:
+      - max-age=31556926; includeSubDomains
+      X-Content-Type-Options:
+      - nosniff
+      X-Frame-Options:
+      - sameorigin
+      X-Ratelimit-Limit:
+      - '20'
+      X-Ratelimit-Remaining:
+      - '19'
+      X-Ratelimit-Reset:
+      - '1647266777'
+      X-Xss-Protection:
+      - 1; mode=block
+    status:
+      code: 200
+      message: OK
+version: 1
