+++ tests/integration/workflows/cassettes/test_workflow_stop_exception.yaml
@@ -1,0 +1,53 @@
+interactions:
+- request:
+    body: '{"force_stop": true}'
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '20'
+      Content-Type:
+      - application/json
+    method: PUT
+    uri: https://reana.cern.ch/api/workflows/4e4e6afc-fd57-4f59-836a-40ae19d2d6d0/status?status=stop
+  response:
+    body:
+      string: "{\n  \"message\": \"Workflow 4e4e6afc-fd57-4f59-836a-40ae19d2d6d0 is\
+        \ not running.\"\n}\n"
+    headers:
+      Access-Control-Allow-Origin:
+      - '*'
+      Access-Control-Expose-Headers:
+      - Content-Type, ETag, Link, X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset
+      Content-Length:
+      - '81'
+      Content-Type:
+      - application/json
+      Date:
+      - Mon, 14 Mar 2022 14:06:17 GMT
+      Referrer-Policy:
+      - strict-origin-when-cross-origin
+      Retry-After:
+      - '1'
+      Strict-Transport-Security:
+      - max-age=31556926; includeSubDomains
+      X-Content-Type-Options:
+      - nosniff
+      X-Frame-Options:
+      - sameorigin
+      X-Ratelimit-Limit:
+      - '20'
+      X-Ratelimit-Remaining:
+      - '19'
+      X-Ratelimit-Reset:
+      - '1647266779'
+      X-Xss-Protection:
+      - 1; mode=block
+    status:
+      code: 409
+      message: Conflict
+version: 1
